fix: Fixed ScoreMessageRequest score to allow decimals

+++ src/main/java/me/loopbreak/hermesanalyzer/objects/request/ScoreMessageRequest.java
@@ -1,4 +1,4 @@
 package me.loopbreak.hermesanalyzer.objects.request;
 
-public record ScoreMessageRequest(int score) {
+public record ScoreMessageRequest(double score) {
 }
